pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-dpo"
version = "0.1.0"
description = "Aligning Video Diffusion Models via VLM Preference Optimization"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.30.0",
    "transformers>=4.40.0",
    "accelerate>=0.28.0",
    "peft>=0.10.0",
    "safetensors>=0.4.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "decord>=0.6.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
eval = [
    "clean-fid>=0.1.35",
    "clip-benchmark>=1.4.0",
    "scipy>=1.11.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]
all = ["vlm-dpo[eval,dev]"]

[project.scripts]
vlm-dpo = "vlm_dpo.cli.main:main"

[tool.setuptools.packages.find]
include = ["vlm_dpo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
